fix: binding issues
